delete duplicated input and fix imports
